Add a thread for init, cleanup, refactor
Move init code to init_and_store
Add an InitAndStoreThread class to be called by init
Allow setting the frequency as a string, instead of an integer. This lets you specifically ask for one of Peter's clock plans, e.g. 25M or a number such as 10000000
Query the frequency back from the sync_role, and use that to label the data
Remove explicit import for Mgt508Ports, now included as part of hapi
Add more debug printing